+++ Assets/Scripts/UI/HUD.cs
@@ -10,9 +10,9 @@ public class HUD : MonoBehaviour
     public Image experienceBar;
     public TextMeshProUGUI level;
 
-    public void UpdateHealthBar(int currentHealth, int maxHealth)
+    public void UpdateHealthBar(float currentHealth, float maxHealth)
     {
-        healthBar.fillAmount = (float)currentHealth / maxHealth;
+        healthBar.fillAmount = currentHealth / maxHealth;
     }
 
     public void UpdateExperienceBar(float currentExperience, float experienceToNextLevel)
